+++ modules/contracts/src.ts/services/ethService.ts
@@ -8,6 +8,8 @@ import {
   IChainServiceStore,
   TransactionReason,
   FullTransferState,
+  UINT_MAX,
+  jsonifyError,
 } from "@connext/vector-types";
 import {
   bufferify,
@@ -26,11 +28,15 @@ import { Wallet } from "@ethersproject/wallet";
 import { BaseLogger } from "pino";
 import PriorityQueue from "p-queue";
 import { AddressZero, HashZero } from "@ethersproject/constants";
+import { parseUnits } from "@ethersproject/units";
 import { MerkleTree } from "merkletreejs";
 
 import { ChannelFactory, VectorChannel } from "../artifacts";
 
 import { EthereumChainReader } from "./ethReader";
+
+export const EXTRA_GAS_PRICE = parseUnits("20", "gwei");
+export const EXTRA_GAS = 50_000;
 
 export class EthereumChainService extends EthereumChainReader implements IVectorChainService {
   private signers: Map<number, Signer> = new Map();
@@ -150,6 +156,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
 
   public async sendDeployChannelTx(
     channelState: FullChannelState,
+    gasPrice: BigNumber,
     deposit?: { amount: string; assetId: string }, // Included IFF createChannelAndDepositAlice
   ): Promise<Result<TransactionResponse, ChainError>> {
     const method = "sendDeployChannelTx";
@@ -187,7 +194,9 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         if (multisigRes.getValue() !== `0x`) {
           return undefined;
         }
-        return channelFactory.createChannel(channelState.alice, channelState.bob);
+        const _gas = await channelFactory.estimateGas.createChannel(channelState.alice, channelState.bob);
+        const gas = _gas.add(EXTRA_GAS);
+        return channelFactory.createChannel(channelState.alice, channelState.bob, { gasPrice, gasLimit: gas });
       });
       if (result.isError) {
         return result as Result<any, ChainError>;
@@ -221,11 +230,23 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         }
         if (multisigRes.getValue() !== `0x`) {
           // multisig deployed, just send deposit
-          return this.sendDepositATx(channelState, amount, AddressZero);
+          return this.sendDepositATx(channelState, amount, AddressZero, gasPrice);
         }
         // otherwise deploy with deposit
+        const _gas = await channelFactory.estimateGas.createChannelAndDepositAlice(
+          channelState.alice,
+          channelState.bob,
+          assetId,
+          amount,
+          {
+            value: amount,
+          },
+        );
+        const gas = _gas.add(EXTRA_GAS);
         return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
           value: amount,
+          gasPrice,
+          gasLimit: gas,
         });
       }) as Promise<Result<TransactionResponse, ChainError>>;
     }
@@ -236,15 +257,19 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       channelState.channelAddress,
       channelState.networkContext.channelFactoryAddress,
       sender,
-      amount,
+      UINT_MAX,
       assetId,
       channelState.networkContext.chainId,
+      gasPrice,
     );
     if (approveRes.isError) {
       return Result.fail(approveRes.getError()!);
     }
     if (approveRes.getValue()) {
       const receipt = await approveRes.getValue()!.wait();
+      if (receipt.status === 0) {
+        return Result.fail(new ChainError(ChainError.reasons.TxReverted, { receipt }));
+      }
       this.log.info({ txHash: receipt.transactionHash, method, assetId }, "Token approval confirmed");
     }
     return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.deployWithDepositAlice, async () => {
@@ -254,9 +279,19 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       }
       if (multisigRes.getValue() !== `0x`) {
         // multisig deployed, just send deposit (will check allowance)
-        return this.sendDepositATx(channelState, amount, assetId);
+        return this.sendDepositATx(channelState, amount, assetId, gasPrice);
       }
-      return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount);
+      const _gas = await channelFactory.estimateGas.createChannelAndDepositAlice(
+        channelState.alice,
+        channelState.bob,
+        assetId,
+        amount,
+      );
+      const gas = _gas.add(EXTRA_GAS);
+      return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
+        gasPrice,
+        gasLimit: gas,
+      });
     }) as Promise<Result<TransactionResponse, ChainError>>;
   }
 
@@ -282,26 +317,28 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       return Result.fail(multisigRes.getError()!);
     }
 
+    const gasPriceRes = await this.getGasPrice(channelState.networkContext.chainId);
+    if (gasPriceRes.isError) {
+      Result.fail(gasPriceRes.getError()!);
+    }
+    const _gasPrice = gasPriceRes.getValue();
+    const gasPrice = _gasPrice.add(EXTRA_GAS_PRICE);
+    this.log.info(
+      {
+        channelAddress: channelState.channelAddress,
+        sender,
+        method,
+        gasPrice: gasPrice.toString(),
+        chainId: channelState.networkContext.chainId,
+      },
+      "Got gas price",
+    );
+
     if (multisigRes.getValue() === `0x`) {
       // Deploy multisig tx
       this.log.info({ channelAddress: channelState.channelAddress, sender, method }, "Deploying channel");
-      const channelFactory = new Contract(
-        channelState.networkContext.channelFactoryAddress,
-        ChannelFactory.abi,
-        signer,
-      );
-      const txRes = await this.sendTxWithRetries(channelState.channelAddress, TransactionReason.deploy, async () => {
-        // If the multisig has been deployed already, do not redeploy
-        const multisigRes = await this.getCode(channelState.channelAddress, channelState.networkContext.chainId);
-        if (multisigRes.isError) {
-          return Result.fail(multisigRes.getError()!);
-        }
-        if (multisigRes.getValue() !== `0x`) {
-          return undefined;
-        }
-        return channelFactory.createChannel(channelState.alice, channelState.bob);
-      });
-      if (txRes.isError) {
+      const txRes = await this.sendDeployChannelTx(channelState, gasPrice);
+      if (txRes.isError && txRes.getError()?.message !== ChainError.reasons.MultisigDeployed) {
         return Result.fail(
           new ChainError(ChainError.reasons.FailedToDeploy, {
             method,
@@ -310,23 +347,15 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           }),
         );
       }
-      const deployTx = txRes.getValue();
+      const deployTx = txRes.isError ? undefined : txRes.getValue();
       if (deployTx) {
         this.log.info({ method, deployTx: deployTx.hash }, "Deploy tx broadcast");
         try {
           this.log.debug("Waiting for event to be emitted");
-          await Promise.race([
-            deployTx.wait(),
-            new Promise((resolve) =>
-              setTimeout(() => {
-                this.log.warn(
-                  { deployTx: deployTx.hash, channel: channelState.channelAddress },
-                  "Did not see event within 15s after tx was mined",
-                );
-                resolve(undefined);
-              }, 15_000),
-            ),
-          ]);
+          const receipt = await deployTx.wait();
+          if (receipt.status === 0) {
+            return Result.fail(new ChainError(ChainError.reasons.TxReverted, { receipt }));
+          }
         } catch (e) {
           this.log.error(`caught ${e.message}`);
           return Result.fail(
@@ -345,9 +374,11 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     }
 
     this.log.info({ sender, method, channel: channelState.channelAddress }, "Sending withdraw tx to chain");
-    return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.withdraw, () =>
-      signer.sendTransaction(minTx),
-    ) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.withdraw, async () => {
+      const _gas = await signer.estimateGas(minTx);
+      const gas = _gas.add(EXTRA_GAS);
+      return signer.sendTransaction({ ...minTx, gasPrice, gasLimit: gas });
+    }) as Promise<Result<TransactionResponse, ChainError>>;
   }
 
   public async sendDepositTx(
@@ -372,6 +403,23 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       return Result.fail(multisigRes.getError()!);
     }
 
+    const gasPriceRes = await this.getGasPrice(channelState.networkContext.chainId);
+    if (gasPriceRes.isError) {
+      Result.fail(gasPriceRes.getError()!);
+    }
+    const _gasPrice = gasPriceRes.getValue();
+    const gasPrice = _gasPrice.add(EXTRA_GAS_PRICE);
+    this.log.info(
+      {
+        channelAddress: channelState.channelAddress,
+        sender,
+        method,
+        gasPrice: gasPrice.toString(),
+        chainId: channelState.networkContext.chainId,
+      },
+      "Got gas price",
+    );
+
     const multisigCode = multisigRes.getValue();
     // alice needs to deploy the multisig
     if (multisigCode === `0x` && sender === channelState.alice) {
@@ -385,7 +433,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         },
         `Deploying channel with deposit`,
       );
-      return this.sendDeployChannelTx(channelState, { amount, assetId });
+      return this.sendDeployChannelTx(channelState, gasPrice, { amount, assetId });
     }
 
     this.log.info({ method, assetId, amount }, "Channel is deployed, sending deposit");
@@ -394,7 +442,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         { method, sender, alice: channelState.alice, bob: channelState.bob },
         "Detected participant A, sending tx",
       );
-      const txRes = await this.sendDepositATx(channelState, amount, assetId);
+      const txRes = await this.sendDepositATx(channelState, amount, assetId, gasPrice);
       if (txRes.isError) {
         this.log.error({ method, error: txRes.getError()?.message }, "Error sending tx");
       } else {
@@ -406,7 +454,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         { method, sender, alice: channelState.alice, bob: channelState.bob },
         "Detected participant B, sendng tx",
       );
-      const txRes = await this.sendDepositBTx(channelState, amount, assetId);
+      const txRes = await this.sendDepositBTx(channelState, amount, assetId, gasPrice);
       if (txRes.isError) {
         this.log.error({ method, error: txRes.getError()?.message }, "Error sending tx");
       } else {
@@ -482,8 +530,18 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         // Register callbacks for saving tx, then return
         response
           .wait() // TODO: confirmation blocks?
-          .then((receipt) => this.store.saveTransactionReceipt(channelAddress, receipt))
-          .catch((e) => this.store.saveTransactionFailure(channelAddress, response.hash, e.message));
+          .then((receipt) => {
+            if (receipt.status === 0) {
+              this.log.error({ method: "sendTxAndParseResponse", receipt }, "Transaction reverted");
+              this.store.saveTransactionFailure(channelAddress, response.hash, "Tx reverted");
+            } else {
+              this.store.saveTransactionReceipt(channelAddress, receipt);
+            }
+          })
+          .catch((e) => {
+            this.log.error({ method: "sendTxAndParseResponse", error: jsonifyError(e) }, "Transaction reverted");
+            this.store.saveTransactionFailure(channelAddress, response.hash, e.message);
+          });
         return response;
       });
       return Result.ok(response);
@@ -504,6 +562,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     amount: string,
     assetId: string,
     chainId: number,
+    gasPrice: BigNumber,
   ): Promise<Result<TransactionResponse | undefined, ChainError>> {
     const signer = this.signers.get(chainId);
     if (!signer?._isSigner) {
@@ -541,7 +600,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       return Result.ok(undefined);
     }
     const approveRes = await this.sendTxWithRetries(channelAddress, TransactionReason.approveTokens, () =>
-      erc20.approve(spender, amount),
+      erc20.approve(spender, amount, gasPrice),
     );
     if (approveRes.isError) {
       this.log.error(
@@ -563,6 +622,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     channelState: FullChannelState,
     amount: string,
     assetId: string,
+    gasPrice: BigNumber,
   ): Promise<Result<TransactionResponse, ChainError>> {
     const signer = this.signers.get(channelState.networkContext.chainId);
     if (!signer?._isSigner) {
@@ -577,9 +637,10 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         channelState.channelAddress,
         channelState.channelAddress,
         channelState.alice,
-        amount,
+        UINT_MAX,
         assetId,
         channelState.networkContext.chainId,
+        gasPrice,
       );
       if (approveRes.isError) {
         this.log.error(
@@ -594,15 +655,18 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       }
       const approveTx = approveRes.getValue();
       if (approveTx) {
-        await approveTx.wait();
+        const receipt = await approveTx.wait();
+        if (receipt.status === 0) {
+          return Result.fail(new ChainError(ChainError.reasons.TxReverted, { receipt }));
+        }
       }
       this.log.info({ txHash: approveTx?.hash, method: "sendDepositATx", assetId }, "Token approval confirmed");
       return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.depositA, () =>
-        vectorChannel.depositAlice(assetId, amount),
+        vectorChannel.depositAlice(assetId, amount, { gasPrice }),
       ) as Promise<Result<TransactionResponse, ChainError>>;
     }
     return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.depositA, () =>
-      vectorChannel.depositAlice(assetId, amount, { value: amount }),
+      vectorChannel.depositAlice(assetId, amount, { value: amount, gasPrice }),
     ) as Promise<Result<TransactionResponse, ChainError>>;
   }
 
@@ -610,6 +674,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     channelState: FullChannelState,
     amount: string,
     assetId: string,
+    gasPrice: BigNumber,
   ): Promise<Result<TransactionResponse, ChainError>> {
     const signer = this.signers.get(channelState.networkContext.chainId);
     if (!signer?._isSigner) {
@@ -623,12 +688,13 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           to: channelState.channelAddress,
           value: BigNumber.from(amount),
           chainId: channelState.networkContext.chainId,
+          gasPrice,
         }),
       ) as Promise<Result<TransactionResponse, ChainError>>;
     } else {
       const erc20 = new Contract(channelState.networkContext.channelFactoryAddress, ERC20Abi, signer);
       return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.depositB, () =>
-        erc20.transfer(channelState.channelAddress, amount),
+        erc20.transfer(channelState.channelAddress, amount, { gasPrice }),
       ) as Promise<Result<TransactionResponse, ChainError>>;
     }
   }