+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,13 @@
 
 ## Next Release
 
+## 0.1.3
+
+- \[browser-node\] Remove from storage if transfer is cancelled.
+- \[browser-node\] Register listener for sender cancellations once
+- \[browser-node\] Don't hang on `crossChainTransfer` if sender-side transfer cancelled
+- \[router\] Handle unverified (sent a single-signed) updates for transfer forwarding on check-in (using `reconcileDeposit` for state consensus)
+
 ## 0.1.3-beta.1
 
 - \[browser-node\] Remove from storage if transfer is cancelled.